backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# Async & Caching
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Filtering & Permissions
django-filter==23.3
django-guardian==2.4.0

# Multi-tenancy (simplified)
django-tenants==3.5.0

# File handling
Pillow==10.3.0

# Utilities
python-decouple==3.8
requests==2.31.0

# API Documentation
drf-spectacular==0.27.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0

# Development
django-extensions==3.2.3
django-debug-toolbar==4.2.0
factory-boy==3.3.0

# Testing
pytest-django==4.7.0
coverage==7.3.2

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring
sentry-sdk==1.38.0
prometheus-client==0.19.0
django-health-check==3.17.0
django-ratelimit==4.1.0
